Adjusted for Bitwise 10 Crypto Index Fund (BITW)
The Function.R has been adapted to generate Portfolios starting daily and running for 1,2 or 3 years ( the Daily function should be use for Crypto use only)
Crypto Portfolios will be setup with 20% for value and 20% for growth.
A new function has been added to fill the gaps in the Crypto 10 Index Fund time series.